requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.2
colorama==0.4.6
idna==3.10
numpy>=2.4.1
pandas>=3.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pytest>=7.0.0
pytest-mock>=3.10.0
requests==2.32.4
six==1.17.0
tzdata==2025.2
urllib3==2.5.0